Refactored Old test files into 1to1 test file to file tested framework, all old tests have been moved
