--- quell-server/__tests__/__oldTests__/updateProtoWithFragment.test.ts
@@ -1,40 +1,0 @@
-import { updateProtoWithFragment } from '../../src/helpers/quellHelpers';
-
-describe('tests for update prototype with fragments on the server side', () => {
-  test('basic prototype object with 2 fields and a fragment, should convert to a protoype with 2 fields and the fields from the fragment without the fragment key on the prototype object', () => {
-    const protoObj = {
-      artists: {
-        __id: null,
-        __args: null,
-        __alias: null,
-        __type: 'artists',
-        id: true,
-        name: true,
-        artistFragment: true,
-      },
-    };
-
-    const fragment = {
-      artistFragment: {
-        instrument: true,
-        band: true,
-        hometown: true,
-      },
-    };
-
-    expect(updateProtoWithFragment(protoObj, fragment)).toEqual({
-
-      artists: {
-        __id: null,
-        __args: null,
-        __alias: null,
-        __type: 'artists',
-        id: true,
-        name: true,
-        instrument: true,
-        band: true,
-        hometown: true
-      },
-    })
-  });
-})
