+++ quell-server/__tests__/helpers/redisHelpers.test.ts
@@ -1,9 +1,99 @@
+import request from 'supertest';
+import app from '../../test-config/test-server';
+import { QuellCache } from '../../src/QuellCache';
+import schema from '../../test-config/testSchema';
+import { getRedisInfo } from '../../src/helpers/redisHelpers';
 
+// tests pass locally, but time out in travis CI build...
 
-describe('Test Suite for redisHelpers.ts', () => {
-    console.log ('Test Suite for redisHelpers.ts');
+// Legacy Test Suite for getRedisInfo from redisHelpers
+describe('server test for getRedisInfo', () => {
+  const Quell = new QuellCache({
+    schema: schema,
+    redisPort: Number(process.env.REDIS_PORT) || 6379,
+    redisHost: process.env.REDIS_HOST || '127.0.0.1',
+    redisPassword: process.env.REDIS_PASSWORD || '',
+  });
 
-    it ('This is a test', () => {
-        console.log ('This is a filler test');
+  app.use(
+    '/redis',
+    ...getRedisInfo({
+      getStats: true,
+      getKeys: true,
+      getValues: true,
     })
-})
+  );
+
+  const server = app.listen(3000, () => {});
+
+  beforeAll(() => {
+    const promise1 = new Promise((resolve, reject) => {
+      resolve(
+        Quell.writeToCache('country--1', {
+          id: '1',
+          capitol: { id: '2', name: 'DC' },
+        })
+      );
+    });
+    const promise2 = new Promise((resolve, reject) => {
+      resolve(Quell.writeToCache('country--2', { id: '2' }));
+    });
+    const promise3 = new Promise((resolve, reject) => {
+      resolve(Quell.writeToCache('country--3', { id: '3' }));
+    });
+    const promise4 = new Promise((resolve, reject) => {
+      resolve(
+        Quell.writeToCache('countries', [
+          'country--1',
+          'country--2',
+          'country--3',
+        ])
+      );
+    });
+    return Promise.all([promise1, promise2, promise3, promise4]);
+  });
+
+  afterAll(() => {
+    server.close();
+    Quell.redisCache.flushAll();
+    Quell.redisCache.quit();
+  });
+
+  it('responds with a 200 status code', async () => {
+    const response = await request(app).get('/redis');
+    expect(response.statusCode).toBe(200);
+  });
+
+  it('gets stats from redis cache', async () => {
+    const response = await request(app).get('/redis');
+    const redisStats = response.body.redisStats;
+    expect(Object.keys(redisStats)).toEqual([
+      'server',
+      'client',
+      'memory',
+      'stats',
+    ]);
+  });
+
+  it('gets keys from redis cache', async () => {
+    const response = await request(app).get('/redis');
+    const redisKeys = response.body.redisKeys;
+    expect(redisKeys).toEqual([
+      'country--2',
+      'country--1',
+      'countries',
+      'country--3',
+    ]);
+  });
+
+  it('gets values from redis cache', async () => {
+    const response = await request(app).get('/redis');
+    const redisValues = response.body.redisValues;
+    expect(redisValues).toEqual([
+      '{"id":"2"}',
+      '{"id":"1","capitol":{"id":"2","name":"DC"}}',
+      '["country--1","country--2","country--3"]',
+      '{"id":"3"}',
+    ]);
+  });
+});
